+++ virtualenvwrapper.sh
@@ -157,6 +157,7 @@ function virtualenvwrapper_run_hook {
             echo "ERROR: virtualenvwrapper_run_hook could not find temporary file $hook_script" 1>&2
             return 2
         fi
+        # cat "$hook_script"
         source "$hook_script"
     fi
     \rm -f "$hook_script" >/dev/null 2>&1
@@ -652,6 +653,162 @@ function cpvirtualenv {
     virtualenvwrapper_run_hook "post_mkvirtualenv"
     virtualenvwrapper_run_hook "post_cpvirtualenv"
 }
+
+#
+# virtualenvwrapper project functions
+#
+
+# Verify that the PROJECT_HOME directory exists
+function virtualenvwrapper_verify_project_home {
+    if [ -z "$PROJECT_HOME" ]
+    then
+        echo "ERROR: Set the PROJECT_HOME shell variable to the name of the directory where projects should be created." >&2
+        return 1
+    fi
+    if [ ! -d "$PROJECT_HOME" ]
+    then
+        [ "$1" != "-q" ] && echo "ERROR: Project directory '$PROJECT_HOME' does not exist.  Create it or set PROJECT_HOME to an existing directory." >&2
+        return 1
+    fi
+    return 0
+}
+
+# Given a virtualenv directory and a project directory,
+# set the virtualenv up to be associated with the 
+# project
+function setvirtualenvproject {
+    typeset venv="$1"
+    typeset prj="$2"
+    if [ -z "$venv" ]
+    then
+        venv="$VIRTUAL_ENV"
+    fi
+    if [ -z "$prj" ]
+    then
+        prj="$(pwd)"
+    fi
+    echo "Setting project for $(basename $venv) to $prj"
+    echo "$prj" > "$venv/.project"
+}
+
+# Show help for mkproject
+function mkproject_help {
+    echo "Usage: mkproject [-t template] [virtualenv options] project_name"
+    echo ""
+    echo "Multiple templates may be selected.  They are applied in the order"
+    echo "specified on the command line."
+    echo
+    echo "Available templates:"
+    echo
+    "$VIRTUALENVWRAPPER_PYTHON" -m virtualenvwrapper.hook_loader -l project.template
+}
+
+# Create a new project directory and its associated virtualenv.
+function mkproject {
+    typeset -a in_args
+    typeset -a out_args
+    typeset -i i
+    typeset tst
+    typeset a
+    typeset t
+    typeset templates
+
+    in_args=( "$@" )
+
+    if [ -n "$ZSH_VERSION" ]
+    then
+        i=1
+        tst="-le"
+    else
+        i=0
+        tst="-lt"
+    fi
+    while [ $i $tst $# ]
+    do
+        a="${in_args[$i]}"
+        # echo "arg $i : $a"
+        case "$a" in
+            -h)
+                echo 'mkproject help:';
+                echo;
+                mkproject_help;
+                echo;
+                echo 'mkvirtualenv help:';
+                echo;
+                mkvirtualenv -h;
+                return;;
+            -t)
+                templates="$templates $a";
+                i=$(( $i + 1 ));;
+            *)
+                if [ ${#out_args} -gt 0 ]
+                then
+                    out_args=( "${out_args[@]-}" "$a" )
+                else
+                    out_args=( "$a" )
+                fi;;
+        esac
+        i=$(( $i + 1 ))
+    done
+
+    set -- "${out_args[@]}"
+
+    # echo "templates $templates"
+    # echo "remainder $@"
+    # return 0
+
+    eval "typeset envname=\$$#"
+    virtualenvwrapper_verify_project_home || return 1
+
+    if [ -d "$PROJECT_HOME/$envname" ]
+    then
+        echo "Project $envname already exists." >&2
+        return 1
+    fi
+
+    mkvirtualenv "$@" || return 1
+
+    cd "$PROJECT_HOME"
+
+    virtualenvwrapper_run_hook project.pre_mkproject $envname
+
+    echo "Creating $PROJECT_HOME/$envname"
+    mkdir -p "$PROJECT_HOME/$envname"
+    setvirtualenvproject "$VIRTUAL_ENV" "$PROJECT_HOME/$envname"
+
+    cd "$PROJECT_HOME/$envname"
+
+    for t in $templates
+    do
+        echo
+        echo "Applying template $t"
+        virtualenvwrapper_run_hook --name $t project.template $envname
+    done
+
+    virtualenvwrapper_run_hook project.post_mkproject
+}
+
+# Change directory to the active project
+function cdproject {
+    virtualenvwrapper_verify_workon_home || return 1
+    virtualenvwrapper_verify_active_environment || return 1
+    if [ -f "$VIRTUAL_ENV/.project" ]
+    then
+        project_dir=$(cat "$VIRTUAL_ENV/.project")
+        if [ ! -z "$project_dir" ]
+        then
+            cd "$project_dir"
+        else
+            echo "Project directory $project_dir does not exist" 1>&2
+            return 1
+        fi
+    else
+        echo "No project set in $VIRTUAL_ENV/.project" 1>&2
+        return 1
+    fi
+    return 0
+}
+
 
 #
 # Invoke the initialization hooks